pyproject.toml
# pyproject.toml

[build-system]
requires      = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nitegame"
version = "0.0.53"
description = "A game engine built on Pygame & ModernGL meant to simplify common game systems, and the process of developing, primarily, 2D games."
readme = "README.md"
authors = [{ name = "Zachery Visger", email = "niteboundx@gmail.com" }]
license = { file = "LICENSE" }
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
]
keywords = ["games", "tools", "engine"]

dependencies = [
    "pygame",
    "numpy",
    "noise",
    "pillow",
    "PyOpenGL",
    "PyOpenGL-accelerate"
]
requires-python = ">=3.9"

[project.urls]
Homepage = "https://github.com/Nitebound/Nitegame-Engine"

[project.scripts]
#realpython = "reader.__main__:main"